Introduce `BoTorchModel` to OSS
Summary:
This PR introduces a simple way to create a custom BoTorch model in Ax.

The new modular `BoTorchModel` object is composed of a choice of a `Surrogate` class and a choice of an `Acquisition` class. The `Surrogate` is a wrapper around a BoTorch GP `Model` and the `Acquisition` is a wrapper around a BoTorch `AcquisitionFunction`. These `BoTorchModel`-s are also storable with JSON and SQL.

Note that the functionality in this commit is under construction and not yet finalized.

A work-in-progress tutorial for the new setup is also part of this commit, as 'tutorials/botorch_modular.ipynb'.

--------

Reviewed By: lena-kashtelyan

Differential Revision: D23376065

fbshipit-source-id: 9d653d3b0bc2f124140192b26965bbc0e5c0f357